sistem pakar tea solutions
Tea Solutions is a system for automatically diagnosing pests on tea plants. Tea solutions is a mobile-based application that can be used by farmers as a supporting material for diagnosing pests on
tea plants as well as providing solutions for handling pests that have been diagnosed. Tea solutions also provides a website that can be used by an expert for pest data management, symptom data,
data handling solutions on tea plants